feat: add pagination support
This PR solves the pagination issue. We don't want to start at 0 every time. In this case, it would be extra work, but in other cases, it may lead to errors, for example, when we want to start not with the first entity and receive it itself - this would be an error.

Now in code we use readdir_simple from NFSFileSystem trait

readdir_simple uses inside itself readir which is used for readdrirplus but we simply take only certain fields for readdir operation.

In future it will be reasonable to separate these functions

Resolves: #32